+++ src/ts/type.ts
@@ -4,7 +4,7 @@
  * A lightweight JavaScript library that attaches a viewer to a specific node type, allowing you to view the CSS properties, attributes, and size/position.
  * 
  * @file        type.ts
- * @version     v1.3.0
+ * @version     v1.4.0
  * @author      Bunoon
  * @license     MIT License
  * @copyright   Bunoon 2024
@@ -17,7 +17,7 @@ export type Position = {
 };
 
 export type Configuration = {
-	cssPropertiesText?: string;
+	cssText?: string;
     attributesText?: string;
     sizeText?: string;
     classesText?: string;
@@ -40,4 +40,5 @@ export type Options = {
     showOnly?: string[] | string;
     allowEditing?: boolean;
     showIdOrNameInTitle?: boolean;
+    showNodeNameInTitle?: boolean;
 };